Sort out the S3TypedStore tests
